server/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose
pydantic